// src/Interfaces/IEvolution.ts
// Generated by https://quicktype.io

export interface IEvolution {
    baby_trigger_item: null;
    chain:             Chain;
    id:                number;
}

export interface Chain {
    evolution_details: any[];
    evolves_to:        any[];
    is_baby:           boolean;
    species:           Species;
}

export interface Species {
    name: string;
    url:  string;
}